Implement mp_stack_top for mp_stack class
Introduced a new mp_stack_top method for mp_stack class to
return the pointer to a top frame of the stack.

Class methods have been redesigned so that this new mp_stack_top
helper fits into the API naturally.

This is required in scope of multikey indexes to keep a pointer
to a multikey frame and extract currently processed item index
of this frame later.

Needed for https://github.com/tarantool/tarantool/issues/1257