Make MP_EXT mp_snprint() and mp_fprint() customizable
MP_EXT in these functions was skipped and printed as 'undefined'.
In Tarantool there are already 3 extensions - decimals, UUID,
errors. It is time to make them being nicely printed too.

The patch makes it possible via virtual MP_EXT serializer, which
by default does the same as before. But in Tarantool is
substituted with correct printer for the mentioned extensions.

Part of https://github.com/tarantool/tarantool/issues/4719

Reviewed-by: Cyrill Gorcunov <gorcunov@gmail.com>
Reviewed-by: Serge Petrenko <sergepetrenko@tarantool.org>
Reviewed-by: Alexander Turenko <alexander.turenko@tarantool.org>