Don't escape forward slash in mp_snprint
mp_snprint formats MsgPack data as a JSON-like string. According to
the JSON format, the forward slash character isn't supposed to be
escaped, but mp_snprint does escape it. Let's fix this.

Note, this should be fine from the Tarantool compatibility point of
view, because Tarantool uses this function only for presenting MsgPack
data in an unspecified human-readable format, e.g. when printing to
the log. If we need to format MsgPack data as valid JSON, we should use
the internal Tarantool JSON encoder which respects the JSON serializer
config and compat module options.

See https://github.com/tarantool/tarantool/issues/6200