Fix mp_read_double_lossy tests freebsd build
Since the mp_read_double_lossy test performs lossy read from the
msgpack, the value it reads is not strictly equal to the one we
compare it with sometimes, which leads to lossy implicit cast on
values equality check.

Fixed this by explicitly casting the value to compare the read
result with to double.